Unbind server port when calling shutdown!
When closing a SelectableChannel, the port isn't immediately unbound. The port is bound until the
channel's Selector runs a selection operation which triggers the deregistration of the
channel from the Selector.

See
https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/nio/channels/Selector.html#selop

Here we only pass in the selector when we're shutting down the event loop because in
other cases we may spawn another channel to handle traffic from that port.

Fixes #4